+++ data-products/src/test/scala/org/sunbird/analytics/job/report/TestCollectionSummaryJobV2.scala
@@ -9,7 +9,7 @@ import org.ekstep.analytics.framework.conf.AppConf
 import org.ekstep.analytics.framework.util.{HadoopFileUtil, JSONUtils}
 import org.ekstep.analytics.framework.{FrameworkContext, JobConfig}
 import org.scalamock.scalatest.MockFactory
-import org.sunbird.analytics.util.UserData
+import org.sunbird.analytics.util.{ReportUtil, UserData}
 
 import scala.collection.mutable
 
@@ -185,7 +185,7 @@ class TestCollectionSummaryJobV2 extends BaseReportSpec with MockFactory {
     val classLoader = getClass.getClassLoader
     val file = new File(classLoader.getResource(resourceName).getFile)
     val absolutePath = file.getAbsolutePath
-    CollectionSummaryJobV2.submitIngestionTask(ingestionServerMockURL, absolutePath)
+    ReportUtil.submitIngestionTask(ingestionServerMockURL, absolutePath)
   }
 
   /*
