+++ data-products/src/test/scala/org/sunbird/analytics/sourcing/TestSourcingSummaryReport.scala
@@ -77,7 +77,7 @@ class TestSourcingSummaryReport extends SparkSpec with Matchers with MockFactory
     val classLoader = getClass.getClassLoader
     val file = new File(classLoader.getResource(resourceName).getFile)
     val absolutePath = file.getAbsolutePath
-    val config = s"""{"search": {"type": "none"},"model": "org.ekstep.analytics.job.report.SourcingReports","modelParams": {"druidIngestionUrl":"https://httpbin.org/post","specPath":"$absolutePath","dbName":"postgres","druidQuery": {"queryType": "groupBy","dataSource": "vdn-content-model-snapshot","intervals": "1901-01-01T00:00:00+00:00/2101-01-01T00:00:00+00:00","aggregations": [{"name": "count","type": "count"}],"dimensions": [{"fieldName": "primaryCategory","aliasName": "primaryCategory"},{"fieldName": "createdBy","aliasName": "createdBy"}],  "filters": [{"type": "equals","dimension": "objectType","value": "Content"}, {"type": "equals","dimension": "sampleContent","value": "false"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"reportConfig": {"id": "funnel_report","metrics": [],"labels": {"reportDate": "Report generation date","visitors": "No. of users opening the project","projectName": "Project Name","initiatedNominations": "No. of initiated nominations","rejectedNominations": "No. of rejected nominations","pendingNominations": "No. of nominations pending review","acceptedNominations": "No. of accepted nominations to the project","noOfContributors": "No. of contributors to the project","noOfContributions": "No. of contributions to the project","pendingContributions": "No. of contributions pending review","approvedContributions": "No. of approved contributions"},"output": [{"type": "csv","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}, {"type": "json","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}]},"store": "azure","format": "csv","key": "druid-reports/","filePath": "druid-reports/","container": "reportPostContainer","folderPrefix": ["slug", "reportName"]},"sparkCassandraConnectionHost": "sunbirdPlatformCassandraHost","druidConfig": {"queryType": "timeseries","dataSource": "telemetry-events-syncts","intervals": "startdate/enddate","aggregations": [{"name": "visitors","type": "count","fieldName": "actor_id"}],"filters": [{"type": "equals","dimension": "context_cdata_id","value": "program_id"}, {"type": "equals","dimension": "edata_pageid","value": "contribution_project_contributions"}, {"type": "equals","dimension": "context_pdata_pid","value": "creation-portal.programs"}, {"type": "equals","dimension": "context_cdata_type","value": "project"}, {"type": "equals","dimension": "context_env","value": "creation-portal"}, {"type": "equals","dimension": "eid","value": "IMPRESSION"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"output": [{"to": "console","params": {"printEvent": false}}],"parallelization": 8,"appName": "Funnel Report Job","deviceMapping": false}""".stripMargin
+    val config = s"""{"search": {"type": "none"},"model": "org.ekstep.analytics.job.report.SourcingReports","modelParams": {"tables": {"programTable": "program", "nominationTable": "nomination"},"druidIngestionUrl":"https://httpbin.org/post","specPath":"$absolutePath","dbName":"postgres","druidQuery": {"queryType": "groupBy","dataSource": "vdn-content-model-snapshot","intervals": "1901-01-01T00:00:00+00:00/2101-01-01T00:00:00+00:00","aggregations": [{"name": "count","type": "count"}],"dimensions": [{"fieldName": "primaryCategory","aliasName": "primaryCategory"},{"fieldName": "createdBy","aliasName": "createdBy"}],  "filters": [{"type": "equals","dimension": "objectType","value": "Content"}, {"type": "equals","dimension": "sampleContent","value": "false"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"reportConfig": {"id": "funnel_report","metrics": [],"labels": {"reportDate": "Report generation date","visitors": "No. of users opening the project","projectName": "Project Name","initiatedNominations": "No. of initiated nominations","rejectedNominations": "No. of rejected nominations","pendingNominations": "No. of nominations pending review","acceptedNominations": "No. of accepted nominations to the project","noOfContributors": "No. of contributors to the project","noOfContributions": "No. of contributions to the project","pendingContributions": "No. of contributions pending review","approvedContributions": "No. of approved contributions"},"output": [{"type": "csv","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}, {"type": "json","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}]},"store": "azure","format": "csv","key": "druid-reports/","filePath": "druid-reports/","container": "reportPostContainer","folderPrefix": ["slug", "reportName"]},"sparkCassandraConnectionHost": "sunbirdPlatformCassandraHost","druidConfig": {"queryType": "timeseries","dataSource": "telemetry-events-syncts","intervals": "startdate/enddate","aggregations": [{"name": "visitors","type": "count","fieldName": "actor_id"}],"filters": [{"type": "equals","dimension": "context_cdata_id","value": "program_id"}, {"type": "equals","dimension": "edata_pageid","value": "contribution_project_contributions"}, {"type": "equals","dimension": "context_pdata_pid","value": "creation-portal.programs"}, {"type": "equals","dimension": "context_cdata_type","value": "project"}, {"type": "equals","dimension": "context_env","value": "creation-portal"}, {"type": "equals","dimension": "eid","value": "IMPRESSION"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"output": [{"to": "console","params": {"printEvent": false}}],"parallelization": 8,"appName": "Funnel Report Job","deviceMapping": false}""".stripMargin
     implicit val jobConfig = JSONUtils.deserialize[JobConfig](config)
 
     implicit val sqlContext = new SQLContext(sc)
@@ -106,7 +106,7 @@ class TestSourcingSummaryReport extends SparkSpec with Matchers with MockFactory
 
     val report = sqlContext.sparkSession.read
       .option("header","false")
-      .json("sourcing/SourcingReport.json")
+      .json("sourcing/SourcingSummaryReport.json")
 
     report.first().getString(0) should be ("0124698765480987654")
     report.first().getString(1) should be ("Practise Question Set")
@@ -117,7 +117,7 @@ class TestSourcingSummaryReport extends SparkSpec with Matchers with MockFactory
   it should "get correct userType for users" in {
     implicit val sc = spark.sparkContext
     implicit val mockFc = mock[FrameworkContext]
-    val config = s"""{"search": {"type": "none"},"model": "org.ekstep.analytics.job.report.SourcingReports","modelParams": {"druidIngestionUrl":"https://httpbin.org/post","specPath":"absolutePath","dbName":"postgres","druidQuery": {"queryType": "groupBy","dataSource": "vdn-content-model-snapshot","intervals": "1901-01-01T00:00:00+00:00/2101-01-01T00:00:00+00:00","aggregations": [{"name": "count","type": "count"}],"dimensions": [{"fieldName": "primaryCategory","aliasName": "primaryCategory"},{"fieldName": "createdBy","aliasName": "createdBy"}],  "filters": [{"type": "equals","dimension": "objectType","value": "Content"}, {"type": "equals","dimension": "sampleContent","value": "false"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"reportConfig": {"id": "funnel_report","metrics": [],"labels": {"reportDate": "Report generation date","visitors": "No. of users opening the project","projectName": "Project Name","initiatedNominations": "No. of initiated nominations","rejectedNominations": "No. of rejected nominations","pendingNominations": "No. of nominations pending review","acceptedNominations": "No. of accepted nominations to the project","noOfContributors": "No. of contributors to the project","noOfContributions": "No. of contributions to the project","pendingContributions": "No. of contributions pending review","approvedContributions": "No. of approved contributions"},"output": [{"type": "csv","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}, {"type": "json","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}]},"store": "azure","format": "csv","key": "druid-reports/","filePath": "druid-reports/","container": "reportPostContainer","folderPrefix": ["slug", "reportName"]},"sparkCassandraConnectionHost": "sunbirdPlatformCassandraHost","druidConfig": {"queryType": "timeseries","dataSource": "telemetry-events-syncts","intervals": "startdate/enddate","aggregations": [{"name": "visitors","type": "count","fieldName": "actor_id"}],"filters": [{"type": "equals","dimension": "context_cdata_id","value": "program_id"}, {"type": "equals","dimension": "edata_pageid","value": "contribution_project_contributions"}, {"type": "equals","dimension": "context_pdata_pid","value": "creation-portal.programs"}, {"type": "equals","dimension": "context_cdata_type","value": "project"}, {"type": "equals","dimension": "context_env","value": "creation-portal"}, {"type": "equals","dimension": "eid","value": "IMPRESSION"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"output": [{"to": "console","params": {"printEvent": false}}],"parallelization": 8,"appName": "Funnel Report Job","deviceMapping": false}""".stripMargin
+    val config = s"""{"search": {"type": "none"},"model": "org.ekstep.analytics.job.report.SourcingReports","modelParams": {"tables": {"programTable": "program", "nominationTable": "nomination"},"druidIngestionUrl":"https://httpbin.org/post","specPath":"absolutePath","dbName":"postgres","druidQuery": {"queryType": "groupBy","dataSource": "vdn-content-model-snapshot","intervals": "1901-01-01T00:00:00+00:00/2101-01-01T00:00:00+00:00","aggregations": [{"name": "count","type": "count"}],"dimensions": [{"fieldName": "primaryCategory","aliasName": "primaryCategory"},{"fieldName": "createdBy","aliasName": "createdBy"}],  "filters": [{"type": "equals","dimension": "objectType","value": "Content"}, {"type": "equals","dimension": "sampleContent","value": "false"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"reportConfig": {"id": "funnel_report","metrics": [],"labels": {"reportDate": "Report generation date","visitors": "No. of users opening the project","projectName": "Project Name","initiatedNominations": "No. of initiated nominations","rejectedNominations": "No. of rejected nominations","pendingNominations": "No. of nominations pending review","acceptedNominations": "No. of accepted nominations to the project","noOfContributors": "No. of contributors to the project","noOfContributions": "No. of contributions to the project","pendingContributions": "No. of contributions pending review","approvedContributions": "No. of approved contributions"},"output": [{"type": "csv","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}, {"type": "json","dims": ["identifier", "channel", "name"],"fileParameters": ["id", "dims"]}]},"store": "azure","format": "csv","key": "druid-reports/","filePath": "druid-reports/","container": "reportPostContainer","folderPrefix": ["slug", "reportName"]},"sparkCassandraConnectionHost": "sunbirdPlatformCassandraHost","druidConfig": {"queryType": "timeseries","dataSource": "telemetry-events-syncts","intervals": "startdate/enddate","aggregations": [{"name": "visitors","type": "count","fieldName": "actor_id"}],"filters": [{"type": "equals","dimension": "context_cdata_id","value": "program_id"}, {"type": "equals","dimension": "edata_pageid","value": "contribution_project_contributions"}, {"type": "equals","dimension": "context_pdata_pid","value": "creation-portal.programs"}, {"type": "equals","dimension": "context_cdata_type","value": "project"}, {"type": "equals","dimension": "context_env","value": "creation-portal"}, {"type": "equals","dimension": "eid","value": "IMPRESSION"}],"postAggregation": [],"descending": "false","limitSpec": {"type": "default","limit": 1000000,"columns": [{"dimension": "count","direction": "descending"}]}},"output": [{"to": "console","params": {"printEvent": false}}],"parallelization": 8,"appName": "Funnel Report Job","deviceMapping": false}""".stripMargin
     implicit val jobConfig = JSONUtils.deserialize[JobConfig](config)
     insertData()
     //mocking for DruidDataFetcher
